# specs/001-gym-class-booking/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Gym Booking API (Draft)
  version: 0.1.0
servers:
  - url: /api
paths:
...
  /class-types:
    get:
      summary: List class types
      responses:
        '200':
          description: OK
    post:
      summary: Create a class type (instructor or admin)
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /class-types/{id}:
    put:
      summary: Update class type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete class type
      responses:
        '204':
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT